extract HealthCertifiedPersonCoordinator from HealthCertificatesTabCoordinator
